render preview in iframe
refs CNVS-24005

This simplifies the style guide so that all examples are pulled
from code blocks in the doc description comment block and are
rendered in an iframe

Now we should be able to render examples for components that need to be
full window height/width (e.g. ReactTray and modal) without having to
make a separate example file.

Note: there is an issue with the hot module replacement throwing errors that
I will try to fix in a future ticket. We can get around it by doing a
page refresh for now.

Change-Id: I8ac040be92471dbbdb06a87083bf0b0437cc9d3b