Make TabList support both controlled and uncontrolled behavior
This change adds a `defaultSelectedIndex` prop to TabList. The new behavior
is as follows:

- When `selectedIndex` is set the component will assume that it should
behave as a "controlled" component and will not store the selected tab in
its own state. An `onChange` prop is required whenever `selectedIndex` is
set otherwise the component will be not change with interaction.

- When the component is "uncontrolled" and it stores the selected tab in
its own state, the `defaultSelectedIndex` will set the selected tab
for the initial render.

Change-Id: I76d80a4f1d588555db99b9322e74d4c35a9b0720
Reviewed-on: https://gerrit.instructure.com/74031
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>