Upgrade React to 15.0.0
Fixed an issue with RangeInput where defaultValue and value were both set
(because of the defaultProp). This was throwing errors with the new React
version (and would have been and issue anyway).

The CSS changes are to make the component display ok when no value is set.

Change-Id: I85eafd0762b3cf77fca5d33f953fde19ce2e1d30
Reviewed-on: https://gerrit.instructure.com/76650
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>