Fix library exports
This commit reworks the webpack loaders so that we don't use the library
exports to decide what should display in the component navigation.

With this change we're also using the library itself for the examples so
if we mess up the exports we should notice right away.

A noteable change is that instead the main component JS code will live in
the index.js file now. The component generator template has been updated to
reflect this change.

Change-Id: Ic70ce8408f2c5d47c0cedcfd4f3774e0248e63f5
Reviewed-on: https://gerrit.instructure.com/76788
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>