Add Tooltip component
Note: I tried to use the react-overlays component but ended up having
to make a lot of changes for a11y and to work with css modules. This
is still using react-overlays for positioning and for the "portal"
functionality.

Change-Id: Ia49eeff8c49d05ea3bf9b32c55868eacc4adc9bc
Reviewed-on: https://gerrit.instructure.com/79864
Tested-by: Jenkins
Reviewed-by: Jared Crystal <jcrystal@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>