Add a11y testing tools
fixes: INSTUI-74

This adds an eslint plugin to check a11y standards in JSX code and
adds react-axe to display a11y errors in component examples in the
developer tools console when in dev mode.

The following components were update to fix lint errors and remove
console warnings in tests:
Checkbox, Image, Progress, RadioInput, RadioInputGroup, TabList

These components should be regression tested for this commit.

To test that the a11y errors are showing in the console, navigate to any
component that is using the 'brand' color (e.g. Link, Tooltip, Button) and
you should see an error that the component doesn't pass the color contrast
rule. This is ignored for these components in the unit tests for the time
being since we're not required to support the 4.5:1 contrast ratio except
in high contrast mode but you can still see it in the console when in dev
mode.

Change-Id: Ibe7efd176703f21a3c3b3e97057737c0282319ce
Reviewed-on: https://gerrit.instructure.com/89772
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>