Fix Transition unmountOnExit prop
fixes INSTUI-162

This fixes the unmounOnExit prop so that it actually removes the contents
of the Transition after the exit transition is complete.

Test plan:
- In a Transition component example in the docs, if unmountOnExit is set
you should be able to inspect using dev tools and see that the Transition
component element is not present in the document.

Change-Id: I8dd4553301f26c140ba8b93dd1c8a54c196de513
Reviewed-on: https://gerrit.instructure.com/94581
Tested-by: Jenkins
Reviewed-by: John Corrigan <jcorrigan@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>