Update modern Checkbox
fixes SNIZZ-143

The largest difference between the modern and canvas checkbox, as far
as I could tell, is that the modern theme uses the brand color for
the label text once checked.

The actual check icon used in Bridge is still different from the one
used in InstUI, and the size is slightly different, but those kinds
of changes require playing with base variables that I'm not sure
design would be ready to commit to.

Change-Id: Ibf77b1925886dd4403e5982ae8f306c7341bb397
Reviewed-on: https://gerrit.instructure.com/98693
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki