Add margin prop to Button and fix sizing/:focus
Fixes: INSTUI-241, INSTUI-403

1) Test plan for button/input sizing:
Paste the following code into an Inst-UI example:

<div>
<TextInput
  isBlock={false}
  label={<ScreenReaderContent>Hi</ScreenReaderContent>} />
<Button>OK</Button>
</div>

Verify that the Button and the TextInput are the same height
in Inspector.

Change the `size` prop of the Button and TextInput to `small`
and then `large`. Verify both components have the same height.

2) Test plan for adding `margin` to Button component:
Go to one of the examples for the Button component, and
paste in <Button margin="xSmall xLarge large medium">Margin!</Button>.

Verify the button now has inline styles for margin applied to it.
Verify the Button component works exactly the same as before.

Change-Id: Ib6e5286c0546cb470c458770e4e998f9c549d3f7
Reviewed-on: https://gerrit.instructure.com/106258
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki
Product-Review: Chris Hart <chart@instructure.com>