fix(Rating): Rating component should be themeable when `animateFill` pro
Themes were not applied on IE 11 for the Rating component when the
animateFill prop is set. The cause was because we're using the
Transition component, the star icons were not present in the document
when the theme was applied. To fix this, we can call this.applyTheme
when the transition begins.

fixes INSTUI-370

Verify that all examples get the a11y colors in the documentation.

Change-Id: I82ed5a9137e2344b41072cc00f22069641308970
Reviewed-on: https://gerrit.instructure.com/107801
Tested-by: Jenkins
Reviewed-by: Pam Hiett <phiett@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>