feat(themeable): Make accessible themes 'immutable'
Prevent overrides to accessible themes via ApplyTheme and theme props.

fixes INSTUI-513

Test plan:
In the ApplyTheme example in the docs, verify that none of the colors
are changed when you switch to either of the -a11y themes.

Note: when you are in a dev env (e.g. npm run start:watch) you should
see console warnings on the ApplyTheme docs page, but if you're in prod
env (e.g. npm start) there should be no warnings.

Change-Id: I7e3385a8c5e50a9213b8851cac9748279112503a