feat(Breadcrumb) Add onClick to BreadcrumbLink
Fixes: INSTUI-560

If an onClick prop is provided to BreadcrumbLink, it should
render a button and respond to the onClick event

Test plan:

In the first example on the Breadcrumb page in INST-UI,
click on the second breadcrumb link (Exploring John Updike).

Verify that the following console log appears: This BreadcrumbLink
was clicked!

Verify that the BreadcrumbLink is a button via Inspector

Change-Id: I1b0bc86dc2ebff51ba508e532d76ec2759231857
Reviewed-on: https://gerrit.instructure.com/114238
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>