feat(FormField, TextInput): Support rtl text for text inputs
Provides support for rtl text using postcss-bidirection plugin syntax.
Specifically for TextInput root classes.
Overrode some of the all: initial styles successfully

This commit is cherry-picked on top of the FormFieldLayout because it depends
on that to be a full solution.

refs INSTUI-505

Test plan:
- Verify that Autocomplete and TextInput fields respect a dir="rtl"
attribute on the <html> tag. Icons should be at the end and
labels/placeholders at the start.
- Set page to rtl by adding a
dir="rtl" attribute to the <html> tag in your inspector.

Change-Id: I51b84e60d493c6f8a3a0e92913d70b14cdfad79b
Reviewed-on: https://gerrit.instructure.com/114553
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>