feat(GridCol): Support rtl text in GridCol
Makes column-offset support rtl text. This is a parent commit to at
least one other commit that is coming (Select).

refs INSTUI-505

Test plan:
- On the Select component page, scroll down to "Select with messages".
The messages should respect text and line up with the select input when
page is set to rtl. Test this in conjunction with the Select commit that is on
top of this one.
- Set page to rtl by adding a dir="rtl" attribute
on the <html> tag in your inspector.

Change-Id: Idbf08bd8765b77bf6733f6e8426a04bd3336f750
Reviewed-on: https://gerrit.instructure.com/114876
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>