build(dependencies): Upgrade node and add a '^' to deps
refs INSTUI-575

This will make it so if the app consuming instructure-ui
has other things that require these same dependencies
but, for example, require a newer version than what we have here
they and this will use the same version of that dependency.
It will make it so the webpack bundle sizes in quizzes
and canvas-lms are smaller.

Test plan:
- Automated tests pass

Change-Id: I12382990992fb50eb47cdf868eff591f12314374
Reviewed-on: https://gerrit.instructure.com/117226
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>
Reviewed-by: Jennifer Stern <jstern@instructure.com>