fix(autocomplete): add min-width to autocomplete
Fixes: INSTUI-714

Test Plan:
- Go to autocomplete and shrink down the screen
- The input arrow no longer disappears, and a set minimum width is
in place to refrain for the input getting too small to use

Note: this min-width overrides widths set explicitly with a prop if
they are smaller than the min-width. A user should not be able to set
a width to be less than what we constrain for the best user experience.

Change-Id: Ic4867079cada00494f576a98682135f574a6da75
Reviewed-on: https://gerrit.instructure.com/122774
Tested-by: Jenkins
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>