fix(FileDrop): make input sibling of label
This is a workaround for a MS Edge bug, which causes the file selection
dialog to keep reappearing after it gets dismissed, making the browser
unusable. The bug has been fixed in recent versions of Edge, but we
still need to support the previous version. The workaround is to make
the file input a sibling of the label element, not a child of it.

Fixes QUIZ-2858

Test plan:

- Use an older version of MS Edge (EdgeHTML < 15.15)
- Click on the FileDrop component
- When the file selection dialog appears, press cancel
- Check that the dialog does not immediately reappear

Change-Id: I3068d4fea5958568b4a287ccbc149e65b1480708
Reviewed-on: https://gerrit.instructure.com/129541
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Chris Hart <chart@instructure.com>