fix(Select): Prevent value change of disabled controlled selects
Fixes INSTUI-937

Test plan:
- Create a controlled Select or navigate to the example in the docs
- Add a disabled prop to the Select
- Try to change the value of the Select with the control
- Verify the value nor the display of the disabled Select changes

The controlled examples in the docs will display a "Value selected: Florida"
type message below the Select that will change even when the Select is disabled.
This value is showing the current state of the example, not that of the
Select itself. Since this particular example isn't actually demonstrating the
disabled functionality, I just assume leave it as is for now.

Change-Id: I666e128176088c41b6f1edaf18bcae64b089ea85
Reviewed-on: https://gerrit.instructure.com/137457
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>