feat(ui-a11y): A default application element
Fixes INSTUI-955

Test Plan:
This is a change that will touch the majority of our
components in ui-overlay. We should test all of them
with supported screen readers to ensure proper behavior.
Pay special attention to:
- Modal
- Tray
- Popover

Using the mount node prop, we can modify the docs
examples to render those components deep inside
the document to make sure the screen reader is still
behaving properly. A good example that is already
configured this way is the Layout component which
can be tested as follows:
- Navigate to the layout component
- Using a supported screen reader, open one of the trays
- You should only be able to navigate within the tray
- Closing the tray, the focus should be returned to the
  trigger, and you should be able to navigate the rest
  of the layout

Change-Id: I276c0e63c208274f303b9df149e1921e7f3d3bcf
Reviewed-on: https://gerrit.instructure.com/140072
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>