docs(ui-docs-client): change tray to open on load
Test plan:
on first load the left navigation tray is open. closing and reopening
work as before.

A solution to the problem of "where are the docs though?" when first
time visitors come to the documentation. The hamburger menu in the upper
left is easily missed.

Change-Id: I43bc4e1db88d20647bc24ec47456b8c71f44d06e
Reviewed-on: https://gerrit.instructure.com/146494
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>