feat(DateInput): add support for disabled days
Fixes INSTUI-1159

Test Plan:

disabledDaysOfWeek:

Create a DateInput component and pass it a
disabledDaysOfWeek array prop (where 0 = Sunday,
1 = Monday, etc). Validate that you cannot
select those days in the date picker. Validate
that it is accessible, and tabbing works correctly
by skipping over the disabled days.

disabledDays array:

Create a DateInput component and pass it an
array of Date objects to the disabledDays
array. Validate that those days are disabled.

disabledDays callback function:

Create a DateInput component and pass it a
callback function to the disabledDays prop.
It will be called for each date and should
return true or false to determine if that
date is disabled.

- Validate that navigating with the arrow
  will skip over disabled dates

All of the above have example components
in the docs.

Change-Id: If49059b5f042c519ce01a1b13763b2a73209c6ac
Reviewed-on: https://gerrit.instructure.com/147327
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>