refactor(ui-docs-client,ui-layout,ui-overlays): Use DrawerLayout in docs app
Fixes: INSTUI-1182, INSTUI-1241

Notes:
This change refactors the docs app to use DrawerLayout and removes
containerQuery from DrawerLayout in favor of addResizeListener.

DrawerTray now has a mountNode prop that can be utilized when it is
overlaying DrawerContent.

DrawerLayout now provides the overlay state to DrawerTray via context
rather than though the overlay prop.

Also in this change, Tooltips and Popovers in docs examples have been
given a mountNode to prevent them from appearing on top of the docs
Nav when it's overlaying the content. Tooltip has been given a
mountNode prop to allow this. This workaround may be removed after
updates are made to DrawerLayout to better handle Popover children.

Finally, Modal's z-index style was moved to resolve an issue in
fullscreen examples.

Test plan:
- Ensure docs app works as expected, with nav open or closed.
- When nav overlays the docs (small screens), ensure you can close the nav
by clicking anywhere else on the page or by pressing ESC.
- When nav does not overlay the docs, ensure clicking the page and
pressing ESC do NOT close the nav.
- Navigate the docs at a small screen size where the nav overlays the
content. Ensure there are no instances of a Popover appearing on top of
the Nav.
- Ensure DrawerLayout still works the same.
- Open fullscreen mode for an example, and ensure the nav doesn't show on
top of the example (INSTUI-1241)

Change-Id: Ie74e429370672ec900fb10136a96c9c9c61c82a4
Reviewed-on: https://gerrit.instructure.com/149528
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>