fix(ui-pagination): fix perf for lots of pages
refs: INSTUI-1332

the <Pagination> compnent would break down and be incredibly slow when
showing pagination for an endpoint with a lot of pages. This fixes that
by doing less work (iteration) all the time (so it is faster for all
existing consumers) and allowing cosumers to parse a sparse array of
children to handle even crazy ammounts of pages.

TEST PLAN:

* go to the docs app and look at the <Pagination> example that says
 "You can handle endpoints that have a lot of pages by..."
* click around to change pages on it, it should be fast
* in fact, if you run in a prod build (where prop type checking is
  turned off), it's perf should be exactly the same as the one that
  renders 9 pages above it

Change-Id: Ife457b338e1baf3b23e8434c73f66e1a34676612
Reviewed-on: https://gerrit.instructure.com/155795
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>