docs(README): add a simple "Quick Start" to README
For newcomers, it's unclear how to get started with instructure-ui 
components and very easy to be perplexed by the requirement to use 
theming.  This patch adds a short `Quick Start` section to the readme 
with a minimal install and code example for developers adopting the 
library.

Change-Id: I9cfe226260356cc1ce340ba69d782bcd09325239
Reviewed-on: https://gerrit.instructure.com/156148
Reviewed-by: Chris Hart <chart@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>
Tested-by: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>