fix(ui-forms): stop JAWS always reading the first option
refs: INSTUI-1304

It looks like the `highlightedIndex` state was not being
set correctly on SelectField, which was causing the
aria-activedescendant attribute to reset to 0 whenever the
menu closed out. The test that might have caught this
was testing for 0, so I updated it to not test for 0 anymore.

TEST PLAN:
- JAWS should now read the selected option when focus is given
  to Select.
- Select should show no regressions from this change in either
  browser or screenreader.

Change-Id: I4daded4ad75ae0f90d7d37cf7c7f8166dd7a559a
Reviewed-on: https://gerrit.instructure.com/156290
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>