feat(ui-presets): Include changelogs in rc packages
This should generate nothing in the case of a full release, but will add
changelogs to the resulting RC package.

This _does_ leave the file tree dirty, but I think that's a feature :)

Note: this also fixes the Y/N confirmation when the release script is
run locally.

Change-Id: I5f6468d1d706186f0adc947d3f18835f9be8c94e
Reviewed-on: https://gerrit.instructure.com/157285
Reviewed-by: Aaron Griffin <agriffin@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins