perf(ui-forms): don't even render options <ul> when not open
before, we were just setting it to display:none, this doesn't even
render it. so it is faster. The biggest thing was the line that did 
`this.optionHeight = el.querySelector('li').clientHeight`
because that would cause the browser to repaint every time

TEST PLAN:

* do a regression test on the <Select> component
* make sure that any css transitions or a11y things still work as they
  should

Change-Id: Icfa007bbe820cb9fdd3f3c7f898b7580ab250871
Reviewed-on: https://gerrit.instructure.com/157949
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Jenkins
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>