fix(ui-forms): prevent scroll jumping with long textareas
refs: INSTUI-1384

When we adjust the dimensions of an autoGrow TextArea,
the browser may temporarily calculate its height to be 0px.
When the TextArea is longer than its scrollable viewport
this can result in the browsers scroll-into-view type
behavior jumping to the top of the viewport.

This change wraps the textarea in a div that will preserve
a minHeight so we can safely adjust the height of the
textarea. We're also removing the debouncing from the
grow method to enable to enable real time measuring of
the textarea.

TEST PLAN:
- Create a TextArea with autoGrow and enough content to
stretch beyond the height of the viewport. Ideally, the
textarea exists at the top of the page for this test.
- Scroll to the bottom of the viewport and type in the
textarea
- Ensure the viewport isn't scrolled to the top.
- Also ensure, the height increases and descreases as it
always has.
- Ensure there are no regressions to TextArea
- Additionally, since we are removing the debounce for
autogrow textareas, it's worth confirming there are no
visible hits to performance as a result of typing
quickly.

Change-Id: I62f05e5206c730b2ce5cea72c69e74d876df318d
Reviewed-on: https://gerrit.instructure.com/157996
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>