fix(docs): add id to README
The docs plugin uses the document id (or the directory name if it's not
specified) to figure out which README should be the home page.

On the docker image the directory name is 'app' so we need to set the
id to match the name field in the repository package.json file.

TEST PLAN:
- no change to the docs with yarn start/yarn start:watch
- the docs should deploy fine in the next release

Change-Id: Ib613d5135408cc314197a78e16ecaf70846e738c
Reviewed-on: https://gerrit.instructure.com/158789
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>