fix(ui-layout): account for stretch positioning in constrain logic
refs: INSTUI-1296

This change fixes the flicker bug seen with Select when it's
cut off by the edge of the browser window. However, it's
actually a problem in calculateElementPosition.js (as always).

In a future commit Select will constrain to it's nearest scroll
parent by default, rather than the window. When that is merged
I expect this issue will occur on scrollable parent elements
instead so we should test that scenario as well.

TEST PLAN:
- View a Select example and resize your browser until it's
partially cut off by the window.
- Expand the dropdown and ensure no flickering occurs.
- Create a select in a horizontally scrollable div and scroll
the div until the Select is partially hidden by the overflow.
- Expand the dropdown and again ensure no flickering occurs.

Change-Id: I4c51bd7e00e460827de80b2bd3981565c7c14ad2
Reviewed-on: https://gerrit.instructure.com/160729
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Ken Meleta <kmeleta@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>