fix(ui-layout): improve vertical margin warning on View
refs: INSTUI-1423

The margin warning in the View component currently fires if
display=auto, as=span, and ANY margin is set to something
other than none or 0. This change updates this logic to
check if top OR bottom margin is set along with the other
two conditions. This prevents the warning from firing when
it shouldn't be.

TEST PLAN:
- Create a View with no properties other than margin.
- Try various combinations of margin values.
- Ensure the warning is only thrown if top or bottom
margin is set to something other than none or 0.
- Also ensure setting only left/right margins do not
throw the warning
- A unit test has been added for this case, and should
pass.

Change-Id: I104511916146a64a308c8647ec2171c6733067a5
Reviewed-on: https://gerrit.instructure.com/163310
Tested-by: Jenkins
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>