docs(ui-overlays): update <Modal> example to reflect proper markup for keyboard-only users
in html forms, you want to wrap the inputs and the submit buttons
inside the same <form> element so that assistive technology knows they
are associated and so that if the user hits enter in any of the form
inputs, it submits the form. you don't want to just put the onSubmit
handler on the onClick of the submit button because there are other ways
to submit a form and you don't want to force those users to tab to that
element just to submit the form.

TEST PLAN:

* load the Modal docs and tab to the input in the modal. 
* hit enter. 
* it should log to the console that the form was submitted and 
  close the form

Change-Id: Ie2f13b09cf14dc67944739fabfe0b70609bda1fd
Reviewed-on: https://gerrit.instructure.com/164720
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Reviewed-by: Stephen Jensen <sejensen@instructure.com>
Product-Review: Stephen Jensen <sejensen@instructure.com>