feat(ui-editable,ui-elements): In place edit
closes INSTUI-1607

Adding Editable and InPlaceEdit components which
provide the infrastructure for in-place editing
data on the page

Also added the DateTime utility functions to the docs app's
globals so they could be used in component examples. (See
the InPlaceEdit's DateTimeInput example.)

Fixed a bug in DateTimeInput.focused, where it didn't take into account
the popup calendar on the DateInput.

Added support for Flex direction=row-reverse and column-reverse.
This facilitated moving the edit button to the start of InPlaceEdit when
right-aligned.

TEST PLAN:
  - the examples in InPlaceEdit/README.md work as expected
    > clicking on the view switches to edit
    > tabbing or clicking out of the edit returns to view with the
      edit button receiving focus. You should be able to keep TABbing
      at this point to move focus down the page.
    > typing ESC while in edit mode should return to view with the edit
      button receiving focus.
    > hovering over the view displays the edit button
    > clicking the edit button switches to edit
    > changing the value in edit is reflected in the view and visa-versa
    > the font styling in the editor is the same as in the view (see the
      Heading sample)
    > if Editable or InPlaceEdit is read-only, the edit button should
      never appear and clicking does not switch from view to edit mode
  - the unit tests pass

Change-Id: Iaa6acdb9041a0e8bde5180119f38f2f67278c1b9
Reviewed-on: https://gerrit.instructure.com/167181
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Mary Anderson <manderson@instructure.com>