feat(ui-test-utils): support Sizzle selectors
This change simplifies the API for the test utils,
so that selectors are now always strings and use the
native supported format with a few additions.

It uses the Sizzle selector engine (used by jquery) to
parse the selector and provide the ability to add
custom selectors.

This also updates ui-testable to provide a static
'selector' string property that can be passed into
the locator util.

Change-Id: I0877e2058f1238dfdad0d6b3b78b66ad057532d5
Reviewed-on: https://gerrit.instructure.com/172128
Tested-by: Jenkins
Reviewed-by: Steve Jensen <sejensen@instructure.com>
Reviewed-by: Ken Meleta <kmeleta@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>