fix(ui-forms): eliminate unnecessary console warnings in Select
refs: INSTUI-1690

The error is logged when a selected option does not correspond
to an option in the dropdown list. The default in this
situation is to fall back to the selected option. In the async
case, we don't have any options, so the selected option never
corresponds to options in the list and the error is always logged.
Defaulting to the selected option in this situation is the correct
behavior. Because this is a supported case and it is behaving
correctly, we should not be logging an error. This commit 
eliminates them.

TEST PLAN:

- Navigate to select
- The error for selected option not corresponding to an option
  in the list should not appear in the console
- Type various options, select various options in both the async
  single select and async multiple select
- No errors for the selected option not corresponding to an
  option in the list should be logged in the console

Change-Id: I3b75a83f4039f3f39f0ec1d376ce964143a3fb32
Reviewed-on: https://gerrit.instructure.com/175346
Tested-by: Jenkins
Reviewed-by: Ken Meleta <kmeleta@instructure.com>
Product-Review: Ken Meleta <kmeleta@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>