feat(ui-test-utils): add a11y test generator util
This adds a generateA11yTests util to the test utils 
library that can be used along with the .examples config for each
component and the axe-core utils to make sure that each
generated example meets accessibility requirements.

generateA11yTests takes 2 arguments: the first is the .examples
file (which is loaded with the component-examples-loader webpack
loader, see the ui-component-examples package), and the second
is an optional argument to run only a specific example (for
use in debugging failures).

e.g. generateA11yTests(ButtonExamples, [1,3]) will run the test
for the 3rd example in the 2nd section based on the examples config.
The section and example indeces are output in the test report to
make it easier to see what those values should be.

Test plan:
yarn test should pass

Change-Id: I7295f0222b2dccf889bd425f8f1430a34274dc52
Reviewed-on: https://gerrit.instructure.com/181561
Tested-by: Jenkins
Reviewed-by: Steve Jensen <sejensen@instructure.com>
QA-Review: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>