feat(ui-postcss-config): minify the css we publish to npm
this will make it so that when we run the release process, were we have NODE_ENV set to production,
it will minify the css it imports into our javascript files we publish to NPM. normally, with our
javascript, we don't worry about minifying it because that is a concern of the consuming app. But
with the css we have to do it ourselves, because in the consuming app all they see is a string and
uglify/babel doesn't know it can minify that.

This commit also shortens the hashes we use in css classnames in prod.
It should be safe because we have both a 5 character “component id”
Plus 4 characters of the modifier hash. Giving us a total of 9 characters
of uniqueness. In fact, that is probably still overkill

TEST PLAN:

* run `yarn bootstrap`
* look at the js files it generated in the "lib" dir of each
  package. the string for the css of themeable components should not have
  whitespace or comments and be minified

to test the css classname hash shortening:
* look at the css string of the generated files in /lib
* it should have class name selectors like:
  .ccJPS_1Gfx instead of .ccJPSvHuev__1Gfxm (note the shorter hashes)

* run `yarn start` and verify examples
* run `yarn start:examples` and verify examples
* pay special attention to IE and Edge and the theme switcher

Change-Id: Iee78983cfe6c50ac8190ff71e1d7029704592554
Reviewed-on: https://gerrit.instructure.com/184494
Tested-by: Jenkins
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Reviewed-by: Pam Hiett <phiett@instructure.com>