fix(ui-menu): always focus the entire menu when opened
refs: INSTUI-1801

Previously, when a menu was first opened, focus would
be placed on the entire menu and in subsequent attempts
to re-open the menu, focus would instead be placed on
the first menu item. This change ensures focus is
always placed on the menu itself when opened.

TEST PLAN:
- in a menu example, open the menu via the trigger
- with the arrow keys, highlight and select a menu item
- re-open the menu and verify focus is given to the
entire menu, not a menu item.
- unit tests should pass

Change-Id: Ibf338113891e333d0524b5f4079cf18b3c19cc2b
Reviewed-on: https://gerrit.instructure.com/185429
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>