fix(ui-forms): prevent select from closing prematurely in IE
refs: INSTUI-1940

In a previous change, we removed a timeout used in
an internal onBlur handler, but IE was relying on
it to track focus between the input and the
options. This change restores that timeout.

TEST PLAN:
- verify selects in internet explorer work as
expected
- verify there are no regressions to select in any
other browsers
- unit tests should pass

Change-Id: Ia7c9a05c85eead8bc9dc545d1e610907c3f8ad0a
Reviewed-on: https://gerrit.instructure.com/187235
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Daniel Sasaki <dsasaki@instructure.com>