fix(ui-utils): allow to prop on elements that allow href
refs: INSTUI-1734, INSTUI-1975, INSTUI-1976

The to prop, for react router links, should be passed
through on elements that already allow an href.

TEST PLAN:
- Verify Links, Buttons, BreadcrumbLinks, and
MenuItems pass the to prop to the dom element
- Verify a warning shows in the console when a `to`
prop is provided without an `as` prop.
- Unit tests should pass

note: When a link or button is rendered with a `to`,
the element may not appear to be clickable and may
not receive clickable styles without rendering as a
React Router Link. This is why we throw a warning.

Change-Id: Ib2f64494f6155d271350c3e38ab8a60a2f00d22d
Reviewed-on: https://gerrit.instructure.com/188775
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Visual-Regression-Test: Ken Meleta <kmeleta@instructure.com>