fix(ui-elements): don’t set type=button on any non-button <Link>s
Fixes: CORE-2942
Refs: CORE-2747

it never makes sense to set type=button on a <span> or <a> or anything
other than an actual <button> or <input type=“button”>.

The problem if we do that comes when people use
https://necolas.github.io/normalize.css/ (or anything derived from it
like canvas-lms does) in their app for their css reset.

normalize.css (and canvas-lms) has these lines
```
/**
 * Correct the inability to style clickable types in iOS and Safari.
 */

button,
[type="button"],
[type="reset"],
[type="submit"] {
  -webkit-appearance: button;
}
```
(https://github.com/necolas/normalize.css/blob/master/normalize.css#L191-L200)

it does it like that, with a bare `[type="button”]` selector, instead of
an `input [type="button”]` selector to keep it’s specificity low, so we
don’t have to do crazy things like we did in g/181185 to get InstUI’s 
styling to take precedence over it.

The problem with that is if <Link> generates markup that looks like
<a href=“whatever” type=“button”> then that will get caught by
that ^ `[type="button”]` selector above and it will make the <a>
half-way look like a button inadvertently.

By never putting type=button on <a>s or anything else besides <button>s,
we never run into that.

And now we can tell apps that ^ kind of reset is ok, and then
revert (or significantly shrink) g/181185 so that we can shrink the size
of <Button>’s css. Because as it is right now, it’s huge (even though
there is not much actually going on in button’s index.js file, it is
the single biggest component in instUI).
See:
https://unpkg.com/@instructure/ui-buttons/es/Button/index.js

So tl;dr: this is all so we can eventually shrink that ^ file size.

TEST PLAN:
* in chrome:
* Add the style `[type="button"] { -webkit-appearance: button}`
  to a page
* then render <Link href=“whatever” onClick={function(){}}>
* it should not have a border and gray background as if it was a button

Change-Id: I982264bb7b51df031c746d5206075879e62ac87d
Reviewed-on: https://gerrit.instructure.com/193235
Tested-by: Jenkins
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Visual-Regression-Test: Jennifer Stern <jstern@instructure.com>